Cast away some compiler warnings.
We know ignored_bytes is > 0 (if it weren't, we'd have treated that as
an error and exited), so just cast it to guint32 to suppress complaints
about a signed vs. unsigned comparison.

Reviewed-on: https://code.wireshark.org/review/20053
(cherry picked from commit cffb18351ff130a5df06e65b87cb9adade74376d)
Reviewed-on: https://code.wireshark.org/review/20054

+++ editcap.c
@@ -582,7 +582,7 @@ is_duplicate(guint8* fd, guint32 len) {
     guint32 new_len;
     guint8 *new_fd;
 
-    if (len <= ignored_bytes) {
+    if (len <= (guint32)ignored_bytes) {
         offset = 0;
     }
 
@@ -624,7 +624,7 @@ is_duplicate_rel_time(guint8* fd, guint32 len, const nstime_t *current) {
     guint32 new_len;
     guint8 *new_fd;
 
-    if (len <= ignored_bytes) {
+    if (len <= (guint32)ignored_bytes) {
         offset = 0;
     }
